Introduce configuration of syslog for Docker TEE image
As of today it only focus on Tuleap & Nginx to introduce
the logic.

There are 2 new environement variables that can be used:
- TULEAP_LOGGER (either `syslog` or not set / empty)
- TULEAP_LOG_SERVER (a given log host server like `myserver:514`)

As of today, you can set `TULEAP_LOGGER` to syslog and observe
the log traffic in /var/log/messages (nginx traffic for instance).

For `TULEAP_LOG_SERVER`, I didn't tested myself as it requires
a log server to test. My goal is to test `tuleap-devel` and our
brand new log server.

nginx configuration was also updated to not log access to assets
in 03-location.conf in order to reduce the amount of logs being
captured and have more meaningful access logs.

go in `plugins/tee_container/docker` and run `docker-compose up`
You need to have nexus account with Read access to docker.tuleap.org
and you need to login on this registry first.

When the stack is up and running, you should go in the web container
and see in /var/log/messages that nginx access and error logs are
there.

Part of request #14864 Centralize logs in syslog & log forwarding

+++ src/tuleap-cfg/Command/Docker/Rsyslog.php
@@ -24,18 +24,43 @@ declare(strict_types=1);
 namespace TuleapCfg\Command\Docker;
 
 use Symfony\Component\Console\Output\OutputInterface;
+use Webimpress\SafeWriter\FileWriter;
 
 final class Rsyslog
 {
+    private const ENV_LOG_SERVER = 'TULEAP_LOG_SERVER';
+
     /**
      * @see https://www.projectatomic.io/blog/2014/09/running-syslog-within-a-docker-container/
      *      https://github.com/rsyslog/rsyslog-docker/blob/master/base/centos7/Dockerfile
      */
-    public function setup(OutputInterface $output): void
+    public function setup(OutputInterface $output, string $tuleap_fqdn): void
     {
         $output->writeln('Setup Rsyslog');
-        unlink('/etc/rsyslog.d/listen.conf');
-        unlink('/etc/rsyslog.conf');
-        copy(__DIR__ . '/../../../../tools/docker/tuleap-aio-c7/rsyslog.conf', '/etc/rsyslog.conf');
+        if (file_exists('/etc/rsyslog.d/listen.conf')) {
+            unlink('/etc/rsyslog.d/listen.conf');
+        }
+        if (file_exists('/etc/rsyslog.conf')) {
+            unlink('/etc/rsyslog.conf');
+        }
+
+        $log_server = getenv(self::ENV_LOG_SERVER);
+        if ($log_server !== false) {
+            copy(__DIR__ . '/../../resources/rsyslog/rsyslog-logforward.conf', '/etc/rsyslog.conf');
+            $logforward = str_replace(
+                [
+                    '%tuleap_fqdn%',
+                    '%log-server%',
+                ],
+                [
+                    $tuleap_fqdn,
+                    $log_server
+                ],
+                file_get_contents(__DIR__ . '/../../resources/rsyslog/logforward.conf')
+            );
+            FileWriter::writeFile('/etc/rsyslog.d/logforward.conf', $logforward, 0644);
+        } else {
+            copy(__DIR__ . '/../../resources/rsyslog/rsyslog.conf', '/etc/rsyslog.conf');
+        }
     }
 }
